refactor: move StandardState and standard wrappers into a single file
